Refactor: Remove static columns from grid system
Static columns being built into the grid system not only created a hacky design system footer, it just felt a little limiting.
This refactor removes it from the grid system and creates an extra component to handle this. It's also been moved into the application level code because at this stage, it only really applies to the design system website rather than to other products/services.